Edit: moved authentication repositories to domain and infra layers
