+++ src/modules/authentication/domain/repositories/IUsersRepository.ts
