+++ src/modules/authentication/infra/repositories/factories/makeUsersRepository.ts
@@ -1,5 +1,5 @@
+import IUsersRepository from 'modules/authentication/domain/repositories/IUsersRepository';
 import UsersRepository from 'modules/authentication/infra/repositories/UsersRepository';
-import IUsersRepository from 'modules/authentication/repositories/IUsersRepository';
 
 export default function makeUsersRepository(): IUsersRepository {
   const usersRepository = new UsersRepository();
