+++ src/modules/authentication/services/LogUserService.ts
@@ -1,6 +1,5 @@
 import IUserLoginDTO from 'modules/authentication/domain/dtos/IUserLoginDTO';
-
-import IUsersRepository from '../repositories/IUsersRepository';
+import IUsersRepository from 'modules/authentication/domain/repositories/IUsersRepository';
 
 interface IExecute {
   token?: string;
