+++ src/modules/authentication/services/factories/makeLogUserService.ts
@@ -1,4 +1,5 @@
-import makeUsersRepository from 'modules/authentication/repositories/factories/makeUsersRepository';
+import makeUsersRepository from 'modules/authentication/infra/repositories/factories/makeUsersRepository';
+
 import LogUserService from '../LogUserService';
 
 export default function makeLogUserService(): LogUserService {
